combine view and component in tests
