--- legacy_tests/Browser/Dirty/Component.php
@@ -1,19 +1,0 @@
-<?php
-
-namespace LegacyTests\Browser\Dirty;
-
-use Illuminate\Support\Facades\View;
-use Livewire\Component as BaseComponent;
-
-class Component extends BaseComponent
-{
-    public $foo = '';
-    public $bar = '';
-    public $baz = '';
-    public $bob = '';
-
-    public function render()
-    {
-        return View::file(__DIR__.'/view.blade.php');
-    }
-}
