--- legacy_tests/Browser/Extensions/Component.php
@@ -1,18 +1,0 @@
-<?php
-
-namespace LegacyTests\Browser\Extensions;
-
-use Illuminate\Support\Facades\View;
-use Livewire\Component as BaseComponent;
-
-class Component extends BaseComponent
-{
-    public $count = 0;
-
-    public function render()
-    {
-        $this->count++;
-
-        return View::file(__DIR__.'/view.blade.php');
-    }
-}
