--- legacy_tests/Browser/Extensions/view.blade.php
@@ -1,7 +1,0 @@
-<div>
-    <button wire:click="$refresh" dusk="refresh">refresh</button>
-
-    @if ($count > 1)
-        <button wire:foo>foo</button>
-    @endif
-</div>
