+++ legacy_tests/Browser/FileDownloads/Test.php
@@ -2,9 +2,9 @@
 
 namespace LegacyTests\Browser\FileDownloads;
 
-use Livewire\Livewire;
 use LegacyTests\Browser\TestCase;
 use Illuminate\Support\Facades\Storage;
+use Livewire\Component;
 
 class Test extends TestCase
 {
@@ -14,7 +14,7 @@ class Test extends TestCase
         $this->onlyRunOnChrome();
 
         $this->browse(function ($browser) {
-            $this->visitLivewireComponent($browser, Component::class)
+            $this->visitLivewireComponent($browser, DownloadComponent::class)
                 ->waitForLivewire()->click('@download')
                 ->waitUsing(5, 75, function () {
                     return Storage::disk('dusk-downloads')->exists('download-target.txt');
@@ -25,7 +25,7 @@ class Test extends TestCase
                 Storage::disk('dusk-downloads')->get('download-target.txt')
             );
 
-            $this->visitLivewireComponent($browser, Component::class)
+            $this->visitLivewireComponent($browser, DownloadComponent::class)
                 ->waitForLivewire()->click('@download-quoted-disposition-filename')
                 ->waitUsing(5, 75, function () {
                     return Storage::disk('dusk-downloads')->exists('download & target.txt');
@@ -39,7 +39,7 @@ class Test extends TestCase
             /**
              * Trigger download with a response return.
              */
-            $this->visitLivewireComponent($browser, Component::class)
+            $this->visitLivewireComponent($browser, DownloadComponent::class)
                 ->waitForLivewire()->click('@download-from-response')
                 ->waitUsing(5, 75, function () {
                     return Storage::disk('dusk-downloads')->exists('download-target2.txt');
@@ -50,7 +50,7 @@ class Test extends TestCase
                 Storage::disk('dusk-downloads')->get('download-target2.txt')
             );
 
-            $this->visitLivewireComponent($browser, Component::class)
+            $this->visitLivewireComponent($browser, DownloadComponent::class)
                 ->waitForLivewire()->click('@download-from-response-quoted-disposition-filename')
                 ->waitUsing(5, 75, function () {
                     return Storage::disk('dusk-downloads')->exists('download & target2.txt');
@@ -71,7 +71,7 @@ class Test extends TestCase
         $this->browse(function ($browser) {
 
             // Download with content-type header.
-            $this->visitLivewireComponent($browser, Component::class)
+            $this->visitLivewireComponent($browser, DownloadComponent::class)
                 ->tap(function ($b) {
                     $b->script([
                         "window.Livewire.hook('commit', ({ component, succeed }) => {
@@ -95,7 +95,7 @@ class Test extends TestCase
             );
 
             // Download with null content-type header.
-            $this->visitLivewireComponent($browser, Component::class)
+            $this->visitLivewireComponent($browser, DownloadComponent::class)
                 ->tap(function ($b) {
                     $b->script([
                         "window.Livewire.hook('commit', ({ component, succeed }) => {
@@ -122,7 +122,7 @@ class Test extends TestCase
              * Download an untitled file with "invalid" content-type header.
              * It mimics this test: dusk="download-an-untitled-file-with-content-type-header"
              */
-            $this->visitLivewireComponent($browser, Component::class)
+            $this->visitLivewireComponent($browser, DownloadComponent::class)
                 ->tap(function ($b) {
                     $b->script([
                         "window.Livewire.hook('commit', ({ component, succeed }) => {
@@ -148,7 +148,7 @@ class Test extends TestCase
             );
 
             // Download an untitled file with content-type header.
-            $this->visitLivewireComponent($browser, Component::class)
+            $this->visitLivewireComponent($browser, DownloadComponent::class)
                 ->tap(function ($b) {
                     $b->script([
                         "window.Livewire.hook('commit', ({ component, succeed }) => {
@@ -175,7 +175,7 @@ class Test extends TestCase
              * Trigger download with a response return.
              */
 
-            $this->visitLivewireComponent($browser, Component::class)
+            $this->visitLivewireComponent($browser, DownloadComponent::class)
                 ->tap(function ($b) {
                     $b->script([
                         "window.Livewire.hook('commit', ({ component, succeed }) => {
@@ -206,7 +206,7 @@ class Test extends TestCase
         $this->onlyRunOnChrome();
 
         $this->browse(function ($browser) {
-            $this->visitLivewireComponent($browser, Component::class)
+            $this->visitLivewireComponent($browser, DownloadComponent::class)
                 ->waitForLivewire()->click('@dispatch-download')
                 ->waitUsing(5, 75, function () {
                     return Storage::disk('dusk-downloads')->exists('download-target.txt');
@@ -217,5 +217,109 @@ class Test extends TestCase
                 Storage::disk('dusk-downloads')->get('download-target.txt')
             );
         });
+    }
+}
+
+class DownloadComponent extends Component
+{
+    protected $listeners = [
+        'download'
+    ];
+
+    public function download()
+    {
+        config()->set('filesystems.disks.dusk-tmp', [
+            'driver' => 'local',
+            'root' => __DIR__,
+        ]);
+
+        return Storage::disk('dusk-tmp')->download('download-target.txt');
+    }
+
+    public function downloadWithContentTypeHeader($contentType = null)
+    {
+        config()->set('filesystems.disks.dusk-tmp', [
+            'driver' => 'local',
+            'root' => __DIR__,
+        ]);
+
+        return Storage::disk('dusk-tmp')->download('download-target.txt', null, ['Content-Type' => $contentType]);
+    }
+
+    public function downloadAnUntitledFileWithContentTypeHeader($contentType = 'text/html')
+    {
+        config()->set('filesystems.disks.dusk-tmp', [
+            'driver' => 'local',
+            'root' => __DIR__,
+        ]);
+
+        return Storage::disk('dusk-tmp')->download('download-target.txt', '', ['Content-Type' => $contentType]);
+    }
+
+    public function downloadFromResponse()
+    {
+        config()->set('filesystems.disks.dusk-tmp', [
+            'driver' => 'local',
+            'root' => __DIR__,
+        ]);
+
+        return response()->download(
+            Storage::disk('dusk-tmp')->path('download-target2.txt')
+        );
+    }
+
+    public function downloadFromResponseWithContentTypeHeader()
+    {
+        config()->set('filesystems.disks.dusk-tmp', [
+            'driver' => 'local',
+            'root' => __DIR__,
+        ]);
+
+        return response()->download(
+            Storage::disk('dusk-tmp')->path('download-target2.txt'),
+            'download-target2.txt',
+            ['Content-Type' => 'text/csv']
+        );
+    }
+
+    public function downloadQuotedContentDispositionFilename()
+    {
+        config()->set('filesystems.disks.dusk-tmp', [
+            'driver' => 'local',
+            'root' => __DIR__,
+        ]);
+
+        return Storage::disk('dusk-tmp')->download('download & target.txt');
+    }
+
+    public function downloadQuotedContentDispositionFilenameFromResponse()
+    {
+        config()->set('filesystems.disks.dusk-tmp', [
+            'driver' => 'local',
+            'root' => __DIR__,
+        ]);
+
+        return response()->download(
+            Storage::disk('dusk-tmp')->path('download & target2.txt')
+        );
+    }
+
+    public function render()
+    {
+        return <<<'HTML'
+            <div>
+                <button wire:click="$dispatch('download')" dusk="dispatch-download">Dispatch Download</button>
+                <button wire:click="download" dusk="download">Download</button>
+                <button wire:click="downloadFromResponse" dusk="download-from-response">Download</button>
+                <button wire:click="downloadQuotedContentDispositionFilename" dusk="download-quoted-disposition-filename">Download</button>
+                <button wire:click="downloadQuotedContentDispositionFilenameFromResponse" dusk="download-from-response-quoted-disposition-filename">Download</button>
+                <button wire:click="downloadWithContentTypeHeader('text/html')" dusk="download-with-content-type-header">Download</button>
+                <button wire:click="downloadWithContentTypeHeader()" dusk="download-with-null-content-type-header">Download</button>
+                <button wire:click="downloadAnUntitledFileWithContentTypeHeader" dusk="download-an-untitled-file-with-content-type-header">Download</button>
+                <button wire:click="downloadAnUntitledFileWithContentTypeHeader('foo')" dusk="download-an-untitled-file-with-invalid-content-type-header">Download</button>
+                <button wire:click="downloadFromResponseWithContentTypeHeader" dusk="download-from-response-with-content-type-header">Download</button>
+                <input dusk="content-type" />
+            </div>
+        HTML;
     }
 }
