--- legacy_tests/Browser/Init/Component.php
@@ -1,21 +1,0 @@
-<?php
-
-namespace LegacyTests\Browser\Init;
-
-use Illuminate\Support\Facades\View;
-use Livewire\Component as BaseComponent;
-
-class Component extends BaseComponent
-{
-    public $output = '';
-
-    public function setOutputToFoo()
-    {
-        $this->output = 'foo';
-    }
-
-    public function render()
-    {
-        return View::file(__DIR__.'/view.blade.php');
-    }
-}
