+++ legacy_tests/Browser/Init/Test.php
@@ -2,21 +2,36 @@
 
 namespace LegacyTests\Browser\Init;
 
+use Livewire\Component;
 use Livewire\Livewire;
-use LegacyTests\Browser\TestCase;
+use Tests\BrowserTestCase;
 
-class Test extends TestCase
+class Test extends BrowserTestCase
 {
-    public function test()
+    public function test_wire_init(): void
     {
-        $this->browse(function ($browser) {
-            $this->visitLivewireComponent($browser, Component::class)
-                /**
-                 * wire:init runs on page load.
-                 */
-                ->waitForText('foo')
-                ->assertSee('foo')
-            ;
-        });
+        Livewire::visit(new class extends Component {
+            public $output = '';
+
+            public function setOutputToFoo()
+            {
+                $this->output = 'foo';
+            }
+
+            public function render()
+            {
+                return <<<'HTML'
+                    <div wire:init="setOutputToFoo">
+                        <span dusk="output">{{ $output }}</span>
+                    </div>
+                HTML;
+            }
+        })
+            /**
+             * wire:init runs on page load.
+             */
+            ->waitForText('foo')
+            ->assertSee('foo')
+        ;
     }
 }
