--- legacy_tests/Browser/Init/view.blade.php
@@ -1,3 +1,0 @@
-<div wire:init="setOutputToFoo">
-    <span dusk="output">{{ $output }}</span>
-</div>
