+++ legacy_tests/Browser/MagicActions/Test.php
@@ -3,40 +3,63 @@
 namespace LegacyTests\Browser\MagicActions;
 
 use Livewire\Livewire;
-use LegacyTests\Browser\TestCase;
-use LegacyTests\Browser\MagicActions\Component;
+use Tests\BrowserTestCase;
 
-class Test extends TestCase
+class Test extends BrowserTestCase
 {
     public function test_magic_toggle_can_toggle_properties()
     {
-        $this->browse(function ($browser) {
-            $this->visitLivewireComponent($browser, Component::class)
-                //Toggle boolean property
-                ->assertSeeIn('@output', 'false')
-                ->waitForLivewire()->click('@toggle')
-                ->assertSeeIn('@output', 'true')
-                ->waitForLivewire()->click('@toggle')
-                ->assertSeeIn('@output', 'false')
+        Livewire::visit(Component::class)
+            //Toggle boolean property
+            ->assertSeeIn('@output', 'false')
+            ->waitForLivewire()->click('@toggle')
+            ->assertSeeIn('@output', 'true')
+            ->waitForLivewire()->click('@toggle')
+            ->assertSeeIn('@output', 'false')
 
-                //Toggle nested boolean property
-                ->assertSeeIn('@outputNested', 'false')
-                ->waitForLivewire()->click('@toggleNested')
-                ->assertSeeIn('@outputNested', 'true')
-                ->waitForLivewire()->click('@toggleNested')
-                ->assertSeeIn('@outputNested', 'false')
-            ;
-        });
+            //Toggle nested boolean property
+            ->assertSeeIn('@outputNested', 'false')
+            ->waitForLivewire()->click('@toggleNested')
+            ->assertSeeIn('@outputNested', 'true')
+            ->waitForLivewire()->click('@toggleNested')
+            ->assertSeeIn('@outputNested', 'false')
+        ;
     }
 
     public function test_magic_event_works()
     {
-        $this->browse(function ($browser) {
-            $this->visitLivewireComponent($browser, Component::class)
-                ->assertDontSeeIn('@outputEvent', 'baz')
-                ->waitForLivewire()->click('@fillBar')
-                ->assertSeeIn('@outputEvent', 'baz')
-            ;
-        });
+        Livewire::visit(Component::class)
+            ->assertDontSeeIn('@outputEvent', 'baz')
+            ->waitForLivewire()->click('@fillBar')
+            ->assertSeeIn('@outputEvent', 'baz')
+        ;
+    }
+}
+
+class Component extends \Livewire\Component
+{
+    public $active = false;
+    public $foo = ['bar' => ['baz' => false]];
+    public $bar = '';
+
+    public function setBar($bar)
+    {
+        $this->bar = $bar;
+    }
+
+    public function render()
+    {
+        return <<<'HTML'
+            <div>
+                <div dusk="output">{{ $active ? "true" : "false" }}</div>
+                <button wire:click="$toggle('active')" dusk="toggle">Toggle Property</button>
+
+                <div dusk="outputNested">{{ $foo['bar']['baz'] ? "true" : "false" }}</div>
+                <button wire:click="$toggle('foo.bar.baz')" dusk="toggleNested">Toggle Nested</button>
+
+                <div dusk="outputEvent">{{ $bar }}</div>
+                <div wire:click="setBar($event.target.id)" id="baz" dusk="fillBar">Click me</div>
+            </div>
+        HTML;
     }
 }
