--- legacy_tests/Browser/Offline/Component.php
@@ -1,14 +1,0 @@
-<?php
-
-namespace LegacyTests\Browser\Offline;
-
-use Illuminate\Support\Facades\View;
-use Livewire\Component as BaseComponent;
-
-class Component extends BaseComponent
-{
-    public function render()
-    {
-        return View::file(__DIR__.'/view.blade.php');
-    }
-}
