+++ legacy_tests/Browser/Offline/Test.php
@@ -2,54 +2,66 @@
 
 namespace LegacyTests\Browser\Offline;
 
-use Laravel\Dusk\Browser;
+use Livewire\Component;
 use Livewire\Livewire;
-use LegacyTests\Browser\TestCase;
+use Tests\BrowserTestCase;
 
-class Test extends TestCase
+class Test extends BrowserTestCase
 {
-    public function test()
+    public function test_wire_offline()
     {
-        $this->browse(function (Browser $browser) {
-            $this->visitLivewireComponent($browser, Component::class)
-                ->assertMissing('@whileOffline')
-                ->offline()
-                ->assertSeeIn('@whileOffline', 'Offline')
-                ->online()
-                ->assertMissing('@whileOffline')
+        Livewire::visit(new class extends Component
+        {
+            public function render()
+            {
+                return <<<'HTML'
+                    <div>
+                        <span wire:offline dusk="whileOffline">Offline</span>
+                        <span wire:offline.class="foo" dusk="addClass"></span>
+                        <span class="hidden" wire:offline.class.remove="hidden" dusk="removeClass"></span>
+                        <span wire:offline.attr="disabled" dusk="withAttribute"></span>
+                        <span wire:offline.attr.remove="disabled" disabled="true" dusk="withoutAttribute"></span>
+                    </div>
+                HTML;
+            }
+        })
+            ->assertMissing('@whileOffline')
+            ->offline()
+            ->assertSeeIn('@whileOffline', 'Offline')
+            ->online()
+            ->assertMissing('@whileOffline')
 
-                /**
-                 * add element class while offline
-                 */
-                ->online()
-                ->assertClassMissing('@addClass', 'foo')
-                ->offline()
-                ->assertHasClass('@addClass', 'foo')
+            /**
+             * add element class while offline
+             */
+            ->online()
+            ->assertClassMissing('@addClass', 'foo')
+            ->offline()
+            ->assertHasClass('@addClass', 'foo')
 
-                /**
-                 * add element class while offline
-                 */
-                ->online()
-                ->assertHasClass('@removeClass', 'hidden')
-                ->offline()
-                ->assertClassMissing('@removeClass', 'hidden')
+            /**
+             * add element class while offline
+             */
+            ->online()
+            ->assertHasClass('@removeClass', 'hidden')
+            ->offline()
+            ->assertClassMissing('@removeClass', 'hidden')
 
-                /**
-                 * add element attribute while offline
-                 */
-                ->online()
-                ->assertAttributeMissing('@withAttribute', 'disabled')
-                ->offline()
-                ->assertAttribute('@withAttribute', 'disabled', 'true')
+            /**
+             * add element attribute while offline
+             */
+            ->online()
+            ->assertAttributeMissing('@withAttribute', 'disabled')
+            ->offline()
+            ->assertAttribute('@withAttribute', 'disabled', 'true')
 
-                /**
-                 * remove element attribute while offline
-                 */
-                ->online()
-                ->assertAttribute('@withoutAttribute', 'disabled', 'true')
-                ->offline()
-                ->assertAttributeMissing('@withoutAttribute', 'disabled')
-            ;
-        });
+            /**
+             * remove element attribute while offline
+             */
+            ->online()
+            ->assertAttribute('@withoutAttribute', 'disabled', 'true')
+            ->offline()
+            ->assertAttributeMissing('@withoutAttribute', 'disabled')
+        ;
     }
 }
