--- legacy_tests/Browser/SupportCollections/view.blade.php
@@ -1,7 +1,0 @@
-<div>
-    <button wire:click="addBar" dusk="add-bar">Add Bar</button>
-
-    @foreach ($things as $thing)
-        <h1>{{ $thing }}</h1>
-    @endforeach
-</div>
