--- legacy_tests/Browser/SupportDateTimes/view.blade.php
@@ -1,7 +1,0 @@
-<div>
-    <span>native-{{ $native->format('m/d/Y') }}</span>
-    <span>nativeImmutable-{{ $nativeImmutable->format('m/d/Y') }}</span>
-    <span>carbon-{{ $carbon->format('m/d/Y') }}</span>
-    <span>carbonImmutable-{{ $carbonImmutable->format('m/d/Y') }}</span>
-    <span>illuminate-{{ $illuminate->format('m/d/Y') }}</span>
-</div>
