--- legacy_tests/Browser/SupportEnums/view.blade.php
@@ -1,3 +1,0 @@
-<div>
-    {{ $enum->value }}
-</div>
