+++ legacy_tests/Browser/SupportStringables/Test.php
@@ -2,16 +2,32 @@
 
 namespace LegacyTests\Browser\SupportStringables;
 
+use Illuminate\Support\Str;
 use Livewire\Livewire;
-use LegacyTests\Browser\TestCase;
+use Tests\TestCase;
 
 class Test extends TestCase
 {
-    public function test()
+    public function test_stringable_support()
     {
-        $this->browse(function ($browser) {
-            $this->visitLivewireComponent($browser, Component::class)
-                ->assertSee('Be excellent to each other');
-        });
+        Livewire::test(new class extends \Livewire\Component {
+            public $string;
+
+            public function mount()
+            {
+                $this->string = Str::of('Be excellent to each other');
+            }
+
+            public function render()
+            {
+                return <<<'HTML'
+                    <div>
+                        {{ $string }}
+                    </div>
+                HTML;
+            }
+        })
+            ->assertSee('Be excellent to each other')
+        ;
     }
 }
