--- legacy_tests/Browser/SupportStringables/view.blade.php
@@ -1,3 +1,0 @@
-<div>
-    {{ $string }}
-</div>
