+++ src/Features/SupportTesting/DuskTestable.php
@@ -3,6 +3,7 @@
 namespace Livewire\Features\SupportTesting;
 
 use Illuminate\Support\Facades\Route;
+use Laravel\Dusk\Browser;
 use function Livewire\{ invade, on };
 use Illuminate\Support\Arr;
 use Tests\TestCase;
@@ -51,6 +52,9 @@ class DuskTestable
         \Laravel\Dusk\Browser::mixin(new DuskBrowserMacros);
     }
 
+    /**
+     * @return Browser
+     */
     static function create($components, $params = [], $queryParams = [])
     {
         if (static::$shortCircuitCreateCall) {
